requirements.txt
fpdf
pdf2jpg
pillow
pytesseract
requests